feat: split tx mapper logic into focused files
